Protected endpoints related to Users

+++ app/blueprints/user/routes.py
@@ -14,6 +14,17 @@ user_bp = Blueprint("users", __name__)
 
 #--------- USER VALIDATION HELPERS ---------
 
+#method to validate a password
+
+def validate_password(password):
+    #rules: min 8 chars, at least 1 uppercase, 1 lowercase, 1 digit, 1 special char
+    pattern = r'^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'
+
+    if re.match(pattern, password):
+        return True
+    
+    return False
+
 #method to validate user update information
 
 def validate_user_update(data):
@@ -51,6 +62,13 @@ def validate_user_update(data):
                 validate_email(data["email_id"])
             except EmailNotValidError as e:
                 errors["email_id"] = str(e)
+                
+    if "password" in data:
+        if not data["password"]:
+            errors["password"] = "Password is required"
+        
+        elif not validate_password(data["password"]):
+            errors["password"] = "password must contain min 8 chars, at least 1 uppercase, 1 lowercase, 1 digit, 1 special char"
         
     return errors
 
@@ -94,7 +112,22 @@ def validate_book_create(data):
 #Get all the users in the database
 
 @user_bp.route("/", methods=["GET"])
+@jwt_required()
 def get_users():
+    
+    #get the user who requested the endpoint
+    
+    requesting_user_id = int(get_jwt_identity())
+    claims = get_jwt()
+    
+    requesting_user = User.query.get(requesting_user_id)
+    
+    if not requesting_user:
+        return jsonify({"Error": "User not Found"}), 404
+    
+    if claims["role"].lower() != "admin":
+        return jsonify({"Error": "Unauthorized Access"}), 403
+    
     users = User.query.all()
     return jsonify([user.to_dict() for user in users])
 
@@ -104,7 +137,7 @@ def get_users():
 @jwt_required()
 def get_user(user_id):
     
-    #get the user who requested this route
+    #get the user who requested the endpoint
     
     requesting_user_id = int(get_jwt_identity())
     claims = get_jwt()
@@ -116,7 +149,7 @@ def get_user(user_id):
     
     #if requesting user is admin then provide the user info
     
-    if requesting_user.role == "admin":
+    if claims["role"].lower() == "admin":
         requested_user = User.query.get(user_id)
         
         if not requested_user:
@@ -136,7 +169,18 @@ def get_user(user_id):
 #Add a book to a user
 
 @user_bp.route("/<int:user_id>/books", methods=["POST"])
+@jwt_required()
 def add_book_to_user(user_id):
+    
+    requesting_user_id = int(get_jwt_identity())
+    
+    requesting_user = User.query.get(requesting_user_id)
+    
+    if not requesting_user:
+        return jsonify({"Error": "User not Found"}), 404
+    
+    if requesting_user_id != user_id:
+        return jsonify({"Error": "Unauthorized Access"}), 403
     
     data = request.get_json()
     errors = validate_book_create(data)
@@ -144,13 +188,8 @@ def add_book_to_user(user_id):
     if errors:
         return jsonify({"Error": "Validation failed", "Details": errors}), 400
     
-    user = User.query.get(user_id)
-    
-    if not user:
-        return jsonify({"Error" : "User Not Found"}), 404
-    
 
-    book = Book(Name=data["Name"], Author=data["Author"], user_id=user.id)
+    book = Book(Name=data["Name"], Author=data["Author"], user_id=user_id)
     db.session.add(book)
     db.session.commit()
     
@@ -160,23 +199,44 @@ def add_book_to_user(user_id):
 #Get all the books owned by a user
 
 @user_bp.route("/<int:user_id>/books", methods=["GET"])
+@jwt_required()
 def get_user_books(user_id):
-    user = User.query.get(user_id)
+    requesting_user_id = int(get_jwt_identity())
+    claims = get_jwt()
     
-    if not user:
+    requesting_user = User.query.get(requesting_user_id)
+    
+    
+    if not requesting_user:
         return jsonify({"Error": "User Not Found"}), 404
     
-    return jsonify([book.to_dict() for book in user.books])
-
+    if claims["role"].lower() == "admin":
+        requested_user = User.query.get(user_id)
+        
+        if not requested_user:
+            return jsonify({"Error": "User Not Found"}), 404
+        
+        return jsonify([book.to_dict() for book in requested_user.books])
+    
+    if requesting_user_id != user_id:
+        return jsonify({"Error": "Unauthorized Access"}), 403
+    
+    return jsonify([book.to_dict() for book in requesting_user.books]), 200
 
 #update user details
 
 @user_bp.route("/<int:user_id>", methods=["PUT"])
+@jwt_required()
 def update_user(user_id):
-    user = User.query.get(user_id)
+    requesting_user_id = int(get_jwt_identity())
     
-    if not user:
+    requesting_user = User.query.get(requesting_user_id)
+    
+    if not requesting_user:
         return jsonify({"Error": "User Not Found"}), 404
+    
+    if requesting_user_id != user_id:
+        return jsonify({"Error": "Unauthorized Access"}), 403
     
     data = request.get_json()
     
@@ -185,24 +245,44 @@ def update_user(user_id):
     if errors:
         return jsonify({"Error": "Validation Failed", "Details": errors}), 400
     
-    user.first_name = data.get("first_name", user.first_name)
-    user.last_name = data.get("last_name", user.last_name)
-    user.email_id = data.get("email_id", user.email_id)
+    requesting_user.first_name = data.get("first_name", requesting_user.first_name)
+    requesting_user.last_name = data.get("last_name", requesting_user.last_name)
+    requesting_user.email_id = data.get("email_id", requesting_user.email_id).lower()
+    if "password" in data:
+        requesting_user.set_password(data["password"])
     
     db.session.commit()
     
-    return jsonify(user.to_dict())
+    return jsonify({"message": "User Details Updated Successfully", "Details": requesting_user.to_dict()}), 200
 
 #Delete a user
 
 @user_bp.route("/<int:user_id>", methods=["DELETE"])
+@jwt_required()
 def delete_user(user_id):
-    user = User.query.get(user_id)
+    requesting_user_id = int(get_jwt_identity())
+    claims = get_jwt()
     
-    if not user:
+    requesting_user = User.query.get(requesting_user_id)
+    
+    if not requesting_user:
         return jsonify({"Error": "User Not Found"}), 404
     
-    db.session.delete(user)
+    if claims["role"].lower() == "admin":
+        user = User.query.get(user_id)
+        
+        if not user:
+            return jsonify({"Error": "User Not Found"}), 404
+        
+        db.session.delete(user)
+        db.session.commit()
+        
+        return jsonify({"Message": f"User with user id {user_id} is deleted successfully"}), 200
+    
+    if requesting_user_id != user_id:
+        return jsonify({"Error": "Unauthorized Access"}), 403
+    
+    db.session.delete(requesting_user)
     db.session.commit()
     
-    return jsonify({"Message": f"User with user id {user_id} is deleted successfully"})
+    return jsonify({"Message": f"User with user id {requesting_user_id} is deleted successfully"}), 200
